Edited function in main.js splitted it into 3 modules
